Eugh, unnecessary casting.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@2554 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/client/keys.c
@@ -724,7 +724,7 @@ void Key_Console (int key)
 		// than the text area to include borders I guess... weird shit.
 		// - Molgrum
 		if (keydown[K_CTRL])
-			con_current->display -= ( ( (int)con_current->vislines - 22 ) / 8 );
+			con_current->display -= ( ( con_current->vislines - 22 ) / 8 );
 		else
 			con_current->display -= 2;
 		
@@ -740,7 +740,7 @@ void Key_Console (int key)
 		// than the text area to include borders I guess... weird shit.
 		// - Molgrum
 		if (keydown[K_CTRL])
-			con_current->display += ( ( (int)con_current->vislines - 22 ) / 8 );
+			con_current->display += ( ( con_current->vislines - 22 ) / 8 );
 		else
 			con_current->display += 2;
